ambassador/requirements.txt
docker-py==1.4.0
requests==2.7.0
six==1.9.0
websocket-client==0.32.0
wheel==0.24.0
